requirements.txt
asgiref==3.7.2
boto3==1.34.55
botocore==1.34.55
chardet==5.2.0
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
jmespath==1.0.1
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
reportlab==4.1.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7